Move LtsTimeStepper implementation to cpp file
This replaces the fake-virtual code with overloads for the MathWrapper
types (hidden in macros), which can be proper virtual functions.
