+++ src/Time/TimeSteppers/AdamsBashforthN.cpp
@@ -4,15 +4,69 @@
 #include "Time/TimeSteppers/AdamsBashforthN.hpp"
 
 #include <algorithm>
+#include <boost/iterator/transform_iterator.hpp>
+#include <iterator>
+#include <limits>
+#include <map>
+#include <ostream>
+#include <pup.h>
+#include <tuple>
 
+#include "NumericalAlgorithms/Interpolation/LagrangePolynomial.hpp"
+#include "Time/BoundaryHistory.hpp"
+#include "Time/EvolutionOrdering.hpp"
 #include "Time/History.hpp"
 #include "Time/SelfStart.hpp"
+#include "Time/Time.hpp"
 #include "Time/TimeStepId.hpp"
+#include "Utilities/CachedFunction.hpp"
 #include "Utilities/EqualWithinRoundoff.hpp"
+#include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/ErrorHandling/Error.hpp"
+#include "Utilities/Gsl.hpp"
 #include "Utilities/Math.hpp"
+#include "Utilities/Overloader.hpp"
 
 namespace TimeSteppers {
+
+namespace {
+// TimeDelta-like interface to a double used for dense output
+struct ApproximateTimeDelta {
+  double delta = std::numeric_limits<double>::signaling_NaN();
+  double value() const { return delta; }
+  bool is_positive() const { return delta > 0.; }
+
+  // Only the operators that are actually used are defined.
+  friend bool operator<(const ApproximateTimeDelta& a,
+                        const ApproximateTimeDelta& b) {
+    return a.value() < b.value();
+  }
+};
+
+// Time-like interface to a double used for dense output
+struct ApproximateTime {
+  double time = std::numeric_limits<double>::signaling_NaN();
+  double value() const { return time; }
+
+  // Only the operators that are actually used are defined.
+  friend ApproximateTimeDelta operator-(const ApproximateTime& a,
+                                        const Time& b) {
+    return {a.value() - b.value()};
+  }
+
+  friend bool operator<(const Time& a, const ApproximateTime& b) {
+    return a.value() < b.value();
+  }
+
+  friend bool operator<(const ApproximateTime& a, const Time& b) {
+    return a.value() < b.value();
+  }
+
+  friend std::ostream& operator<<(std::ostream& s, const ApproximateTime& t) {
+    return s << t.value();
+  }
+};
+}  // namespace
 
 AdamsBashforthN::AdamsBashforthN(const size_t order) : order_(order) {
   if (order_ < 1 or order_ > maximum_order) {
@@ -133,7 +187,7 @@ std::vector<double> AdamsBashforthN::constant_coefficients(const size_t order) {
 }
 
 void AdamsBashforthN::pup(PUP::er& p) {
-  LtsTimeStepper::Inherit::pup(p);
+  LtsTimeStepper::pup(p);
   p | order_;
 }
 
@@ -225,6 +279,310 @@ bool AdamsBashforthN::can_change_step_size_impl(
            std::is_sorted(history.begin(), history.end(), less)));
 }
 
+template <typename T>
+void AdamsBashforthN::add_boundary_delta_impl(
+    const gsl::not_null<T*> result,
+    const TimeSteppers::BoundaryHistoryEvaluator<T>& coupling,
+    const TimeSteppers::BoundaryHistoryCleaner& cleaner,
+    const TimeDelta& time_step) const {
+  const auto signed_order =
+      static_cast<typename decltype(cleaner.local_end())::difference_type>(
+          cleaner.integration_order());
+
+  ASSERT(cleaner.local_size() >= cleaner.integration_order(),
+         "Insufficient data to take an order-" << cleaner.integration_order()
+         << " step.  Have " << cleaner.local_size() << " times, need "
+         << cleaner.integration_order());
+  cleaner.local_mark_unneeded(cleaner.local_end() - signed_order);
+
+  if (std::equal(cleaner.local_begin(), cleaner.local_end(),
+                 cleaner.remote_end() - signed_order)) {
+    // GTS
+    ASSERT(cleaner.remote_size() >= cleaner.integration_order(),
+           "Insufficient data to take an order-" << cleaner.integration_order()
+           << " step.  Have " << cleaner.remote_size() << " times, need "
+           << cleaner.integration_order());
+    cleaner.remote_mark_unneeded(cleaner.remote_end() - signed_order);
+  } else {
+    const auto remote_step_for_step_start =
+        std::upper_bound(cleaner.remote_begin(), cleaner.remote_end(),
+                         *(cleaner.local_end() - 1),
+                         evolution_less<Time>{time_step.is_positive()});
+    ASSERT(remote_step_for_step_start - cleaner.remote_begin() >= signed_order,
+           "Insufficient data to take an order-" << cleaner.integration_order()
+           << " step.  Have "
+           << remote_step_for_step_start - cleaner.remote_begin()
+           << " times before the step, need " << cleaner.integration_order());
+    cleaner.remote_mark_unneeded(remote_step_for_step_start - signed_order);
+  }
+
+  boundary_impl(result, coupling, *(cleaner.local_end() - 1) + time_step);
+}
+
+template <typename T>
+void AdamsBashforthN::boundary_dense_output_impl(
+    const gsl::not_null<T*> result,
+    const TimeSteppers::BoundaryHistoryEvaluator<T>& coupling,
+    const double time) const {
+  return boundary_impl(result, coupling, ApproximateTime{time});
+}
+
+template <typename T, typename TimeType>
+void AdamsBashforthN::boundary_impl(const gsl::not_null<T*> result,
+                                    const BoundaryHistoryEvaluator<T>& coupling,
+                                    const TimeType& end_time) const {
+  // Might be different from order_ during self-start.
+  const auto current_order = coupling.integration_order();
+
+  ASSERT(current_order <= order_,
+         "Local history is too long for target order (" << current_order
+         << " should not exceed " << order_ << ")");
+  ASSERT(coupling.remote_size() >= current_order,
+         "Remote history is too short (" << coupling.remote_size()
+         << " should be at least " << current_order << ")");
+
+  // Avoid billions of casts
+  const auto order_s = static_cast<
+      typename BoundaryHistoryEvaluator<T>::iterator::difference_type>(
+      current_order);
+
+  // Start and end of the step we are trying to take
+  const Time start_time = *(coupling.local_end() - 1);
+  const auto time_step = end_time - start_time;
+
+  // We define the local_begin and remote_begin variables as the start
+  // of the part of the history relevant to this calculation.
+  // Boundary history cleanup happens immediately before the step, but
+  // boundary dense output happens before that, so there may be data
+  // left over that was needed for the previous step and has not been
+  // cleaned out yet.
+  const auto local_begin = coupling.local_end() - order_s;
+
+  if (std::equal(local_begin, coupling.local_end(),
+                 coupling.remote_end() - order_s)) {
+    // No local time-stepping going on.
+    const auto coefficients =
+        get_coefficients(local_begin, coupling.local_end(), time_step);
+
+    auto local_it = local_begin;
+    auto remote_it = coupling.remote_end() - order_s;
+    for (auto coefficients_it = coefficients.rbegin();
+         coefficients_it != coefficients.rend();
+         ++coefficients_it, ++local_it, ++remote_it) {
+      *result +=
+          time_step.value() * *coefficients_it * *coupling(local_it, remote_it);
+    }
+    return;
+  }
+
+  ASSERT(current_order == order_,
+         "Cannot perform local time-stepping while self-starting.");
+
+  const evolution_less<> less{time_step.is_positive()};
+  const auto remote_begin =
+      std::upper_bound(coupling.remote_begin(), coupling.remote_end(),
+                       start_time, less) -
+      order_s;
+
+  ASSERT(std::is_sorted(local_begin, coupling.local_end(), less),
+         "Local history not in order");
+  ASSERT(std::is_sorted(remote_begin, coupling.remote_end(), less),
+         "Remote history not in order");
+  ASSERT(not less(start_time, *(remote_begin + (order_s - 1))),
+         "Remote history does not extend far enough back");
+  ASSERT(less(*(coupling.remote_end() - 1), end_time),
+         "Please supply only older data: " << *(coupling.remote_end() - 1)
+         << " is not before " << end_time);
+
+  // Union of times of all step boundaries on any side.
+  const auto union_times = [&coupling, &local_begin, &remote_begin, &less]() {
+    std::vector<Time> ret;
+    ret.reserve(coupling.local_size() + coupling.remote_size());
+    std::set_union(local_begin, coupling.local_end(), remote_begin,
+                   coupling.remote_end(), std::back_inserter(ret), less);
+    return ret;
+  }();
+
+  using UnionIter = typename decltype(union_times)::const_iterator;
+
+  // Find the union times iterator for a given time.
+  const auto union_step = [&union_times, &less](const Time& t) {
+    return std::lower_bound(union_times.cbegin(), union_times.cend(), t, less);
+  };
+
+  // The union time index for the step start.
+  const auto union_step_start = union_step(start_time);
+
+  // min(union_times.end(), it + order_s) except being careful not
+  // to create out-of-range iterators.
+  const auto advance_within_step = [order_s,
+                                    &union_times](const UnionIter& it) {
+    return union_times.end() - it >
+                   static_cast<typename decltype(union_times)::difference_type>(
+                       order_s)
+               ? it + static_cast<typename decltype(
+                          union_times)::difference_type>(order_s)
+               : union_times.end();
+  };
+
+  // Calculating the Adams-Bashforth coefficients is somewhat
+  // expensive, so we cache them.  ab_coefs(it, step) returns the
+  // coefficients used to step from *it to *it + step.
+  auto ab_coefs = make_overloader(
+      make_cached_function<std::tuple<UnionIter, TimeDelta>, std::map>(
+          [order_s](const std::tuple<UnionIter, TimeDelta>& args) {
+            return get_coefficients(
+                std::get<0>(args) -
+                    static_cast<typename UnionIter::difference_type>(order_s -
+                                                                     1),
+                std::get<0>(args) + 1, std::get<1>(args));
+          }),
+      make_cached_function<std::tuple<UnionIter, ApproximateTimeDelta>,
+                           std::map>(
+          [order_s](const std::tuple<UnionIter, ApproximateTimeDelta>& args) {
+            return get_coefficients(
+                std::get<0>(args) -
+                    static_cast<typename UnionIter::difference_type>(order_s -
+                                                                     1),
+                std::get<0>(args) + 1, std::get<1>(args));
+          }));
+
+  // The value of the coefficient of `evaluation_step` when doing
+  // a standard Adams-Bashforth integration over the union times
+  // from `step` to `step + 1`.
+  const auto base_summand = [&ab_coefs, &end_time, &union_times](
+                                const UnionIter& step,
+                                const UnionIter& evaluation_step) {
+    if (step + 1 != union_times.end()) {
+      const TimeDelta step_size = *(step + 1) - *step;
+      return step_size.value() *
+             ab_coefs(std::make_tuple(
+                 step, step_size))[static_cast<size_t>(step - evaluation_step)];
+    } else {
+      const auto step_size = end_time - *step;
+      return step_size.value() *
+             ab_coefs(std::make_tuple(
+                 step, step_size))[static_cast<size_t>(step - evaluation_step)];
+    }
+  };
+
+  for (auto local_evaluation_step = local_begin;
+       local_evaluation_step != coupling.local_end();
+       ++local_evaluation_step) {
+    const auto union_local_evaluation_step = union_step(*local_evaluation_step);
+    for (auto remote_evaluation_step = remote_begin;
+         remote_evaluation_step != coupling.remote_end();
+         ++remote_evaluation_step) {
+      double deriv_coef = 0.;
+
+      if (*local_evaluation_step == *remote_evaluation_step) {
+        // The two elements stepped at the same time.  This gives a
+        // standard Adams-Bashforth contribution to each segment
+        // making up the current step.
+        const auto union_step_upper_bound =
+            advance_within_step(union_local_evaluation_step);
+        for (auto step = union_step_start;
+             step < union_step_upper_bound;
+             ++step) {
+          deriv_coef += base_summand(step, union_local_evaluation_step);
+        }
+      } else {
+        // In this block we consider a coupling evaluation that is not
+        // performed at equal times on the two sides of the mortar.
+
+        // Makes an iterator with a map to give time as a double.
+        const auto make_lagrange_iterator = [](const auto& it) {
+          return boost::make_transform_iterator(
+              it, [](const Time& t) { return t.value(); });
+        };
+
+        const auto union_remote_evaluation_step =
+            union_step(*remote_evaluation_step);
+        const auto union_step_lower_bound =
+            std::max(union_step_start, union_remote_evaluation_step);
+
+        // Compute the contribution to an interpolation over the local
+        // times to `remote_evaluation_step->value()`, which we will
+        // use as the coupling value for that time.  If there is an
+        // actual evaluation at that time then skip this because the
+        // Lagrange polynomial will be zero.
+        if (not std::binary_search(local_begin, coupling.local_end(),
+                                   *remote_evaluation_step, less)) {
+          const auto union_step_upper_bound =
+              advance_within_step(union_remote_evaluation_step);
+          for (auto step = union_step_lower_bound;
+               step < union_step_upper_bound;
+               ++step) {
+            deriv_coef += base_summand(step, union_remote_evaluation_step);
+          }
+          deriv_coef *=
+              lagrange_polynomial(make_lagrange_iterator(local_evaluation_step),
+                                  remote_evaluation_step->value(),
+                                  make_lagrange_iterator(local_begin),
+                                  make_lagrange_iterator(coupling.local_end()));
+        }
+
+        // Same qualitative calculation as the previous block, but
+        // interpolating over the remote times.  This case is somewhat
+        // more complicated because the latest remote time that can be
+        // used varies for the different segments making up the step.
+        if (not std::binary_search(remote_begin, coupling.remote_end(),
+                                   *local_evaluation_step, less)) {
+          auto union_step_upper_bound =
+              advance_within_step(union_local_evaluation_step);
+          if (coupling.remote_end() - remote_evaluation_step > order_s) {
+            union_step_upper_bound = std::min(
+                union_step_upper_bound,
+                union_step(*(remote_evaluation_step + order_s)));
+          }
+
+          auto control_points = make_lagrange_iterator(
+              remote_evaluation_step - remote_begin >= order_s
+                  ? remote_evaluation_step - (order_s - 1)
+                  : remote_begin);
+          for (auto step = union_step_lower_bound;
+               step < union_step_upper_bound;
+               ++step, ++control_points) {
+            deriv_coef +=
+                base_summand(step, union_local_evaluation_step) *
+                lagrange_polynomial(
+                    make_lagrange_iterator(remote_evaluation_step),
+                    local_evaluation_step->value(), control_points,
+                    control_points +
+                        static_cast<typename decltype(
+                            control_points)::difference_type>(order_s));
+          }
+        }
+      }
+
+      if (deriv_coef != 0.) {
+        // Skip the (potentially expensive) coupling calculation if
+        // the coefficient is zero.
+        *result += deriv_coef *
+                   *coupling(local_evaluation_step, remote_evaluation_step);
+      }
+    }  // for remote_evaluation_step
+  }  // for local_evaluation_step
+}
+
+template <typename Iterator, typename Delta>
+std::vector<double> AdamsBashforthN::get_coefficients(
+    const Iterator& times_begin, const Iterator& times_end, const Delta& step) {
+  if (times_begin == times_end) {
+    return {};
+  }
+  std::vector<double> steps;
+  // This may be slightly more space than we need, but we can't get
+  // the exact amount without iterating through the iterators, which
+  // is not necessarily cheap depending on the iterator type.
+  steps.reserve(maximum_order);
+  for (auto t = times_begin; std::next(t) != times_end; ++t) {
+    steps.push_back((*std::next(t) - *t).value());
+  }
+  steps.push_back(step.value());
+  return get_coefficients_impl(steps);
+}
+
 bool operator==(const AdamsBashforthN& lhs, const AdamsBashforthN& rhs) {
   return lhs.order_ == rhs.order_;
 }
@@ -234,6 +592,7 @@ bool operator!=(const AdamsBashforthN& lhs, const AdamsBashforthN& rhs) {
 }
 
 TIME_STEPPER_DEFINE_OVERLOADS(AdamsBashforthN)
+LTS_TIME_STEPPER_DEFINE_OVERLOADS(AdamsBashforthN)
 }  // namespace TimeSteppers
 
 PUP::able::PUP_ID TimeSteppers::AdamsBashforthN::my_PUP_ID =  // NOLINT
