+++ src/Time/TimeSteppers/TimeStepper.hpp
@@ -11,19 +11,15 @@
 #include "DataStructures/DataBox/PrefixHelpers.hpp"
 #include "DataStructures/MathWrapper.hpp"
 #include "Parallel/CharmPupable.hpp"
+#include "Time/BoundaryHistory.hpp"
 #include "Time/History.hpp"
 #include "Time/TimeStepId.hpp"
-#include "Utilities/FakeVirtual.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/TMPL.hpp"
 
 /// \cond
 class TimeDelta;
-namespace TimeSteppers {
-template <typename LocalVars, typename RemoteVars, typename CouplingResult>
-class BoundaryHistory;
-}  // namespace TimeSteppers
 /// \endcond
 
 /// \ingroup TimeSteppersGroup
@@ -292,21 +288,21 @@ class TimeStepper : public PUP::able {
 // LtsTimeStepper -> TimeStepper -> AdamsBashforthN -> LtsTimeStepper
 // include loop causes AdamsBashforthN to be defined before its base
 // class if LtsTimeStepper is included first.
-namespace LtsTimeStepper_detail {
-DEFINE_FAKE_VIRTUAL(boundary_dense_output)
-DEFINE_FAKE_VIRTUAL(add_boundary_delta)
-}  // namespace LtsTimeStepper_detail
 
 /// \ingroup TimeSteppersGroup
 ///
 /// Base class for TimeSteppers with local time-stepping support,
 /// derived from TimeStepper.
+///
+/// Several of the member functions of this class are templated and
+/// perform type erasure before forwarding their arguments to the
+/// derived classes.  This is implemented using the macros \ref
+/// LTS_TIME_STEPPER_DECLARE_OVERLOADS, which must be placed in a
+/// private section of the class body, and
+/// LTS_TIME_STEPPER_DEFINE_OVERLOADS(derived_class), which must be
+/// placed in the cpp file.
 class LtsTimeStepper : public TimeStepper {
  public:
-  using Inherit =
-      LtsTimeStepper_detail::FakeVirtualInherit_boundary_dense_output<
-          LtsTimeStepper_detail::FakeVirtualInherit_add_boundary_delta<
-              LtsTimeStepper>>;
   // When you add a class here, remember to add it to TimeStepper as well.
   using creatable_classes = tmpl::list<TimeSteppers::AdamsBashforthN>;
 
@@ -326,6 +322,25 @@ class LtsTimeStepper : public TimeStepper {
   using BoundaryReturn = db::unprefix_variables<
       std::result_of_t<const Coupling&(LocalVars, RemoteVars)>>;
 
+/// \cond
+#define LTS_TIME_STEPPER_DECLARE_VIRTUALS_IMPL(_, data)                       \
+  virtual void add_boundary_delta_forward(                                    \
+      gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,                 \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const TimeSteppers::BoundaryHistoryCleaner& cleaner,                    \
+      const TimeDelta& time_step) const = 0;                                  \
+  virtual void boundary_dense_output_forward(                                 \
+      gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,                 \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const double time) const = 0;
+
+  GENERATE_INSTANTIATIONS(LTS_TIME_STEPPER_DECLARE_VIRTUALS_IMPL,
+                          (MATH_WRAPPER_TYPES))
+#undef LTS_TIME_STEPPER_DECLARE_VIRTUALS_IMPL
+  /// \endcond
+
   /// \brief Compute the change in a boundary quantity due to the
   /// coupling on the interface.
   ///
@@ -334,6 +349,17 @@ class LtsTimeStepper : public TimeStepper {
   /// quantity.  These values may be used to form a linear combination
   /// internally, so the result should have appropriate mathematical
   /// operators defined to allow that.
+  ///
+  /// Derived classes must implement this as a function with signature
+  ///
+  /// ```
+  /// template <typename T>
+  /// void add_boundary_delta_impl(
+  ///     gsl::not_null<T*> result,
+  ///     const TimeSteppers::BoundaryHistoryEvaluator<T>& coupling,
+  ///     const TimeSteppers::BoundaryHistoryCleaner& cleaner,
+  ///     const TimeDelta& time_step) const;
+  /// ```
   ///
   /// \note
   /// Unlike the `update_u` methods, which overwrite the `result`
@@ -345,18 +371,28 @@ class LtsTimeStepper : public TimeStepper {
       const gsl::not_null<BoundaryHistoryType<LocalVars, RemoteVars, Coupling>*>
           history,
       const TimeDelta& time_step, const Coupling& coupling) const {
-    return LtsTimeStepper_detail::fake_virtual_add_boundary_delta<
-        creatable_classes>(this, result, history, time_step, coupling);
+    return add_boundary_delta_forward(&*make_math_wrapper(result),
+                                      history->evaluator(coupling),
+                                      history->cleaner(), time_step);
   }
 
+  /// Derived classes must implement this as a function with signature
+  ///
+  /// ```
+  /// template <typename T>
+  /// void boundary_dense_output_impl(
+  ///     gsl::not_null<T*> result,
+  ///     const TimeSteppers::BoundaryHistoryEvaluator<T>& coupling,
+  ///     const double time) const;
+  /// ```
   template <typename LocalVars, typename RemoteVars, typename Coupling>
   void boundary_dense_output(
       const gsl::not_null<BoundaryReturn<LocalVars, RemoteVars, Coupling>*>
           result,
       const BoundaryHistoryType<LocalVars, RemoteVars, Coupling>& history,
       const double time, const Coupling& coupling) const {
-    return LtsTimeStepper_detail::fake_virtual_boundary_dense_output<
-        creatable_classes>(this, result, history, time, coupling);
+    return boundary_dense_output_forward(&*make_math_wrapper(result),
+                                         history.evaluator(coupling), time);
   }
 
   /// Substep LTS integrators are not supported, so this is always 1.
@@ -370,6 +406,54 @@ class LtsTimeStepper : public TimeStepper {
     return next_time_id(current_id, time_step);
   }
 };
+
+/// \cond
+#define LTS_TIME_STEPPER_DECLARE_OVERLOADS_IMPL(_, data)                      \
+  void add_boundary_delta_forward(                                            \
+      gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,                 \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const TimeSteppers::BoundaryHistoryCleaner& cleaner,                    \
+      const TimeDelta& time_step) const override;                             \
+  void boundary_dense_output_forward(                                         \
+      gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,                 \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const double time) const override;
+
+#define LTS_TIME_STEPPER_DEFINE_OVERLOADS_IMPL(_, data)                       \
+  void TIME_STEPPER_DERIVED_CLASS(data)::add_boundary_delta_forward(          \
+      const gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,           \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const TimeSteppers::BoundaryHistoryCleaner& cleaner,                    \
+      const TimeDelta& time_step) const {                                     \
+    return add_boundary_delta_impl(result, coupling, cleaner, time_step);     \
+  }                                                                           \
+  void TIME_STEPPER_DERIVED_CLASS(data)::boundary_dense_output_forward(       \
+      const gsl::not_null<TIME_STEPPER_WRAPPED_TYPE(data)*> result,           \
+      const TimeSteppers::BoundaryHistoryEvaluator<TIME_STEPPER_WRAPPED_TYPE( \
+          data)>& coupling,                                                   \
+      const double time) const {                                              \
+    return boundary_dense_output_impl(result, coupling, time);                \
+  }
+/// \endcond
+
+/// \ingroup TimeSteppersGroup
+/// Macro declaring overloaded detail methods in classes derived from
+/// TimeStepper.  Must be placed in a private section of the class
+/// body.
+#define LTS_TIME_STEPPER_DECLARE_OVERLOADS                         \
+  GENERATE_INSTANTIATIONS(LTS_TIME_STEPPER_DECLARE_OVERLOADS_IMPL, \
+                          (MATH_WRAPPER_TYPES))
+
+/// \ingroup TimeSteppersGroup
+/// Macro defining overloaded detail methods in classes derived from
+/// TimeStepper.  Must be placed in the cpp file for the derived
+/// class.
+#define LTS_TIME_STEPPER_DEFINE_OVERLOADS(derived_class)          \
+  GENERATE_INSTANTIATIONS(LTS_TIME_STEPPER_DEFINE_OVERLOADS_IMPL, \
+                          (MATH_WRAPPER_TYPES), (derived_class))
 
 
 #include "Time/TimeSteppers/AdamsBashforthN.hpp"  // IWYU pragma: keep
